Revert "SF: Add support for 64 bit consumer usage flag"
This reverts commit 6897f770125fa616584c29c3a3382d6064eb0d3b.

+++ services/surfaceflinger/DisplayUtils.h
@@ -58,7 +58,7 @@ class DisplayUtils {
             sp<IGraphicBufferProducer> &producer, sp<IGraphicBufferProducer> bqProducer,
             sp<IGraphicBufferConsumer> bqConsumer, String8 currentStateDisplayName,
             bool currentStateIsSecure);
-    bool canAllocateHwcDisplayIdForVDS(uint64_t usage);
+    bool canAllocateHwcDisplayIdForVDS(int usage);
     bool skipColorLayer(const char* layerType);
     DisplayUtils();
   private:
